[FLINK-4840] [metrics] Measure latency/delay of record processing and expose it as a metric
